+++ examples/transformers_text.py
@@ -203,7 +203,7 @@ if not args.finetune:
     kwargs = {
         "text_stype":
         text_stype,
-        "text_embedder_cfg":
+        "col_to_text_embedder_cfg":
         TextEmbedderConfig(text_embedder=text_encoder, batch_size=10),
     }
 else: