+++ torch_frame/datasets/multimodal_text_benchmark.py
@@ -451,11 +451,15 @@ class MultimodalTextBenchmark(torch_frame.data.Dataset):
             df[df[target_col] < 0][target_col] = 0
         return df
 
-    def __init__(self, root: str, name: str,
-                 text_stype: torch_frame.stype = torch_frame.text_embedded,
-                 col_to_text_embedder_cfg: Optional[Union[Dict[
-                     str, TextEmbedderConfig], TextEmbedderConfig]] = None,
-                 text_tokenizer_cfg: Optional[TextTokenizerConfig] = None):
+    def __init__(
+        self,
+        root: str,
+        name: str,
+        text_stype: torch_frame.stype = torch_frame.text_embedded,
+        col_to_text_embedder_cfg: Optional[Union[Dict[str, TextEmbedderConfig],
+                                                 TextEmbedderConfig]] = None,
+        text_tokenizer_cfg: Optional[TextTokenizerConfig] = None,
+    ):
         assert name in self.classification_datasets | self.regression_datasets
         self.root = root
         self.name = name